pyproject.toml
[tool.poetry]
name = "todo-app-api"
version = "0.1.0"
description = ""
authors = ["uuuta"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
boto3 = "^1.35.87"
mangum = "^0.19.0"
pydantic-settings = "^2.7.0"
pydantic = "^2.10.4"
pydantic-core = "^2.27.2"
tzdata = "^2024.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
flake8 = "^7.1.1"
black = "^24.10.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s"
testpaths = ["tests"]
pythonpath = ["app"]
